+++ scripts/run_cpp_example.py
@@ -98,7 +98,7 @@ GUI_SCENE_EXAMPLE_DEFAULT_ARGS = {
     "wam_ikfast": ("--scene", "wam-ikfast"),
     "fetch": (),
     "tinkertoy": ("--scene", "tinkertoy"),
-    "drag_and_drop": ("--scene", "drag-and-drop"),
+    "drag_and_drop": (),
     "empty": ("--scene", "drag-and-drop"),
     "simple_frames": (),
     "soft_bodies": ("--scene", "soft-bodies"),